server,storage/engine: enforce rocksdb max open file limit
Previously, if we had a large-enough open file ulimit, we'd configure
RocksDB without an open file limit (`max_open_files = -1`). This
configuration enabled a fast path where RocksDB could avoid an lookup in
its open files cache, since it could assume that every opened file would
stay open forever.

Unfortunately, a 2TB restore creates tens of thousands of 2MB SSTs,
quickly exceeding the open file ulimit of 10k. With `max_open_files = -1`
the call to `open` that exceeds the limit crashes the server.

Instead, set the `max_open_files` to the configured ulimit so that
RocksDB evicts a file that hasn't been recently accessed instead of
crashing the server. This introduces two performance hits: an obvious
hit, where accessing restored data is slow because the necessary files
have been evicted from the open files cache, and a less-obvious hit,
where every operation now requires an open-file cache lookup, even when
the number of SSTs is well below the open file limit.

+++ pkg/server/config.go
@@ -325,13 +325,13 @@ func GetTotalMemory(ctx context.Context) (int64, error) {
 // On Windows there is no need to change the file descriptor, known as handles,
 // limit. This limit cannot be changed and is approximately 16,711,680. See
 // https://blogs.technet.microsoft.com/markrussinovich/2009/09/29/pushing-the-limits-of-windows-handles/
-func setOpenFileLimit(physicalStoreCount int) (int, error) {
+func setOpenFileLimit(physicalStoreCount int) (uint64, error) {
 	return setOpenFileLimitInner(physicalStoreCount)
 }
 
 // SetOpenFileLimitForOneStore sets the soft limit for open file descriptors
 // when there is only one store.
-func SetOpenFileLimitForOneStore() (int, error) {
+func SetOpenFileLimitForOneStore() (uint64, error) {
 	return setOpenFileLimit(1)
 }
